requirements.txt
torch>=2.2
numpy>=1.24
tqdm>=4.66
matplotlib>=3.7
pyyaml>=6.0
